requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
python-docx>=0.8.11
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
Pillow>=10.0.0
requests>=2.31.0
pydantic>=2.0.0
typing-extensions>=4.7.0

# OCR Dependencies
easyocr>=1.7.0
opencv-python>=4.8.0
google-cloud-vision>=3.4.0
